+++ spec/files/worker_init/first_file.rb
@@ -1,0 +1,15 @@
+class FirstFile
+  def initialize
+    # do nothing in this example
+  end
+
+  def run(a_file, *args)
+    puts args
+
+    args
+  end
+
+  def finalize
+    # do nothing
+  end
+end